requirements.txt
bedrock-agentcore
strands-agents
websockets
asyncio
boto3
python-dotenv
